requirements.txt
pytest-playwright
pytest
pytest-html
httpx
pytest-xdist